Fix a bug and add a test for pip in subprocesses (#53)
Previously, if an async task started `ensure_has_pip()`, another call to
`ensure_has_pip()` would see that `DID_INIT_PIP` was set to true and
would immediately return (instead of waiting for the first task to
complete).

This PR fixes that issue by using `tokio::sync::OnceCell`. It also adds
a test that checks the pip version in a subprocess.