Trigger Cirrus CI builds from the nightly GitHub Action (#80)
This PR changes the GH actions nightly build to trigger a Cirrus CI
build (for `aarch64` mac and linux). Once that build is complete, the
action downloads the artifacts and reuploads them as GH actions
artifacts.

This lets us have all the artifacts in one place. It also lets us
control and trigger nightly builds from one place (GH Actions)

We can add another step in the future that actually releases the nightly
artifacts for all the architectures we support

### Test Plan

Tested by manually triggering the workflow on another branch