Replace Cirrus CI and GH Actions linux builds with Buildkite builds (#114)
This PR replaces the Cirrus CI `aarch64` Linux build with a Buildkite
build that has agents running on AWS. The Cirrus CI Linux workers have
been consistently failing to start for a while ("instance got
rescheduled").

We also replace the GitHub actions `x86` Linux build with a Buildkite
build. Unfortunately, we don't get enough disk space on GH actions for
testing with LLMs. Since we're responsible for running the CI agents
that are used with Buildkite, we can allocate more disk space.