Correctly handle Lunchbox symlink paths (#140)
The paths returned by `read_link` in Lunchbox are now always absolute
paths (i.e. relative to the filesystem root).

This PR updates `OverlayFS` to handle this. It also updates the loader
in `carton-runner-py` and tests to ensure that symlinks work correctly.

See https://github.com/VivekPanyam/lunchbox/pull/12 and
https://github.com/VivekPanyam/ZipFS/pull/2 for more details.

### Test plan

Updated a test in `carton-runner-py` to do more in-depth testing of
symlinks within python models.