Remove generic storage parameters from `Tensor` (#167)
This PR removes the generic storage parameter from `Tensor<Storage>`.
This lets us delete a lot of code and generally cleans up the `Tensor`
interface.

The way it does this is by creating a `GenericTensorStorage` struct that
can efficiently wrap any implementation of the `TypedStorage` trait. It
can do this type erasure without any dynamic dispatch overhead during
the life of a tensor (i.e. until destruction). This does not have any
measurable runtime performance cost according to the benchmarks I did.

This is an interface breaking change, but since we're at a `0.0.x`
version, [it's
okay](https://doc.rust-lang.org/cargo/reference/semver.html).

## Details

`GenericTensorStorage` stores a view of the underlying data alongside a
trait object of the `TypedStorage` implementation that actually owns the
data. This ensures that the underlying data does not become deallocated
while the `GenericTensorStorage` instance exists.

Importantly, we never actually use the trait object. This lets us avoid
any dynamic dispatch overhead from this change during the life of the
tensor. When the tensor is dropped, the underlying `TypedStorage` impl
is dropped as well (which may have overhead, but this is insignificant).

This process does add a heap allocation with `Box`, but it only slows
down tensor creation + destruction by ~50 nanoseconds in microbenchmarks
on my machine. In slightly more practical benchmarks, the difference is
not measurable.

## Test Plan

CI + local testing of the Python bindings