Initial implementation of C++ bindings (#174)
This PR implements C++ bindings for Carton on top of the C bindings.

It does not yet provide access to metadata or packing, but it does
support asynchronously loading a model, asynchronously running
inference, and interacting with tensors.

See the included `source/carton-bindings-cpp/README.md` along with
examples in `source/carton-bindings-cpp/tests/`. The `README` provides
some rationale for why the C++ bindings are implemented on top of the C
bindings rather than directly on top of Rust code using something like
`cxx`.

### Test plan

Added tests in C++ along with a Rust test that compiles and runs all of
the C++ tests (i.e. `cargo test` runs all the C++ binding tests as well)